requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
_tflow_select==2.3.0
absl-py==0.12.0
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.4.2
anyio==3.6.2
apscheduler==3.9.1.post1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports-zoneinfo==0.2.1
beautifulsoup4==4.11.1
blas==1.0
blinker==1.4
boto3==1.26.89
botocore==1.29.89
brotlipy==0.7.0
c-ares==1.19.0
ca-certificates==2023.01.10
cachetools==4.2.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==7.0
colorama==0.4.6
croniter==1.3.8
crypto==1.4.1
cryptography==39.0.2
cycler==0.11.0
dnspython==2.3.0
elasticsearch==7.10.0
faker==15.3.4
fastapi==0.95.2
fftw==3.3.9
flask-apscheduler==1.12.1
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-marshmallow==0.13.0
flask-migrate==2.5.3
flask-mysqldb==0.2.0
flask-sqlalchemy==2.4.1
flask==1.1.1
flatbuffers==1.12.0
fonttools==4.38.0
freetype==2.12.1
frozenlist==1.3.3
gast==0.3.3
gevent==21.8.0
google-auth-oauthlib==0.4.3
google-auth==2.6.0
google-pasta==0.2.0
google==3.0.0
greenlet==1.1.3.post0
grpcio==1.42.0
gunicorn==20.1.0
h11==0.14.0
h5py==2.10.0
halo==0.0.31
hdf5==1.10.6
idna==2.8
imageio==2.28.1
importlib-metadata==4.11.3
inflect==6.0.4
intel-openmp==2021.4.0
itsdangerous==1.1.0
jaydebeapi==1.2.3
jinja2==2.10.3
jmespath==1.0.1
joblib==1.2.0
jpeg==9e
jpype1==1.3.0
json-logging==1.0.5
keras-base==2.4.3
keras-preprocessing==1.1.2
keras==2.4.3
kiwisolver==1.4.4
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libdeflate==1.8
libffi==3.3
libgcc-ng==11.2.0
libgfortran-ng==11.2.0
libgfortran5==11.2.0
libgomp==11.2.0
libpng==1.6.39
libprotobuf==3.20.3
libstdcxx-ng==11.2.0
libtiff==4.4.0
libwebp-base==1.2.4
log-symbols==0.0.14
lz4-c==1.9.4
mako==1.1.2
markdown==3.4.1
markupsafe==1.1.1
marshmallow-sqlalchemy==0.23.1
marshmallow==3.8.0
matplotlib==3.5.3
mkl-service==2.4.0
mkl==2021.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
multidict==6.0.2
mysql-connector-python==8.0.31
mysqlclient==1.4.5
naked==0.1.32
ncurses==6.3
networkx==2.6.3
numpy-base==1.21.5
numpy==1.21.6
oauthlib==3.2.1
olefile==0.46
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openssl==1.1.1t
opt_einsum==3.3.0
packaging==23.0
pandas==1.3.5
parquet-cli==1.3
parquet-tools==0.2.0
pendulum==2.1.2
pillow==9.5.0
pip==22.2.2
prettytable==3.7.0
protobuf==3.20.1
pwinput==1.0.3
pyarrow==11.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.10.7
pyenchant==3.2.2
pyjwt==1.7.1
pymongo==4.3.3
pyopenssl==23.0.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python==3.7.15
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
pytzdata==2020.1
pywavelets==1.3.0
pyyaml==6.0
readline==8.2
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==67.6.0
shellescape==3.8.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spinners==0.0.24
sqlacodegen==2.3.0.post1
sqlalchemy==1.3.11
sqlite==3.39.3
starlette==0.27.0
tabulate==0.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
tensorflow-base==2.3.0
tensorflow-estimator==2.6.0
tensorflow==2.3.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tk==0.1.0
typing-extensions==4.3.0
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2022.6
tzlocal==4.2
unidecode==1.1.1
urllib3==1.25.7
uvicorn==0.22.0
wcwidth==0.2.6
werkzeug==0.16.0
wheel==0.37.1
wrapt==1.12.1
xz==5.2.6
yaml==0.2.5
yarl==1.8.1
zipp==3.11.0
zlib==1.2.13
zope-event==4.5.0
zope-interface==5.5.2
zstd==1.5.2
